Create proper game logic and tests
Update the class to represent a simple Battleship game.
The field setup is done in a method. Potentially could be extracted
as a part of UI offering customisable experience. Game is handled by
setting the target coordinates and then calling the tick() method.
Check for isRunning() as the game loop condition.